A partial fix to use the good MAIL FROM directive when we transmit an email
Currently, the MAIL FROM used for the incoming email is transfered as is
which can lead to an SPF error when when we send back the email to the
real destination under our identity. This patch replace the MAIL FROM by
the user under our identity to let the receiver to check correctly the
SPF record from our identity.

Note that the verifier (in front of our infra) already wrote a
Received-SPF of the real sender and already check the SPF of it. The
real receiver can have a trace of the SPF result of the real sender even
if we replaced it by our identity.